huggingface_deployment/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# Data validation and serialization
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Document processing
python-docx>=0.8.11
reportlab>=4.0.0

# Task queue and background jobs
celery>=5.3.0

# OpenAI integration
openai>=1.0.0

# Hugging Face integration
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Optional: For Redis backend (if using Redis with Celery)
# redis>=5.0.0

# Optional: For PostgreSQL (if switching back to PostgreSQL)
# psycopg2-binary>=2.9.0
# asyncpg>=0.29.0 